requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
streamlit>=1.28.0

